+++ ShopModel/PurchaseList.cs
@@ -15,7 +15,7 @@ namespace ShopModel
         public string seller_phone { set; get; }
         public DateTime date { set; get; }
         public int goods_num { set; get; }
-        public float unit_price { set; get; }
-        public float total_price { set; get; }
+        public decimal unit_price { set; get; }
+        public decimal total_price { set; get; }
     }
 }
